Move BasicContent ObjectManager into Context
