+++ Context/EzContext.php
@@ -24,7 +24,6 @@ class EzContext implements KernelAwareContext
     use Object\User;
     use Object\Role;
     use Object\FieldType;
-    use Object\BasicContent;
 
     /**
      * @var \Symfony\Component\HttpKernel\KernelInterface
